Refactoring : 여러 수정사안들 수정 및 리팩토링
