+++ frontend/pages/room/createroom/roomsetting.js
@@ -25,14 +25,6 @@ const RoomSetting = () => {
     getMap();
   }, []);
 
-  useEffect(() => {
-    console.log(roomTitle);
-    console.log(exercise);
-    console.log(limitPeopleCount);
-    console.log(roomArea.area);
-    console.log(roomArea.areaDetail);
-  });
-
   const getMap = () => {
     const mapScript = document.createElement("script");
     mapScript.async = true;
