+++ frontend/pages/room/createroom/roomtaginfo.js
@@ -4,6 +4,11 @@ import { useSelector } from "react-redux";
 import { postCreateRoom } from "../../../api/rooms";
 import { FailResponse } from "../../../api/failResponse";
 import RoomInfoNavBar from "../../../components/roomInfoNavBar";
+import SetRoomImages from "../../../components/rooms/setRoomImages";
+
+/* 수정 필요 */
+// 1. 완료 버튼 클릭 시, 새로 만들어진 방으로 이동
+// 2. SetRoomImages 컴포넌트에서 roomImage 데이터 가져오기 (진행중)
 
 const RoomTagInfo = () => {
   const roomInfo = useSelector((state) => state.createRoomReducer);
@@ -12,17 +17,33 @@ const RoomTagInfo = () => {
   const [roomContent, setRoomContent] = useState("");
 
   // 방 이미지
-  const [image, setImage] = useState("");
-  const [imagesrc, setImagesrc] = useState("");
-  const [roomImage, setRoomImage] = useState([]);
-  const [imagePreview, setImagePreview] = useState([]);
+  const [roomImages, setRoomImages] = useState([]);
 
   // 방 태그
   const [tag, setTag] = useState([]);
+  const tagsAge = ["10대", "20대", "30대", "40대", "50대", "60대", "70대"];
+  const tagsLevel = ["입문만", "초보만", "중수만", "고수만"];
+  const tagsGender = ["남자만", "여자만"];
 
-  /* 수정 필요 */
-  // 1. 완료 버튼 클릭 시, 새로 만들어진 방으로 이동
-  // 2. 대표사진에 대한 구현
+  // 태그 선택 함수
+  const onClickTag = (e) => {
+    if (e.target.classList[2] === "tag-clicked") {
+      e.target.classList.remove("tag-clicked");
+      setTag((prev) =>
+        prev.filter((el) => {
+          return el !== e.target.innerText;
+        })
+      );
+    } else {
+      if (tag.length > 9) {
+        e.preventDefault();
+        alert("태그는 최대 10개 선택 가능합니다!");
+      } else {
+        e.target.classList.add("tag-clicked");
+        setTag((prev) => [...prev, e.target.innerText]);
+      }
+    }
+  };
 
   // 예외 처리 및 서버에 방 생성 요청
   const callCreateRoomRequest = (e) => {
@@ -54,118 +75,6 @@ const RoomTagInfo = () => {
     }
   };
 
-  // 이미지 source 인코딩
-  const encodeFileToBase64 = async (file) => {
-    return new Promise((resolve) => {
-      const reader = new FileReader();
-      const baseURL = "";
-      reader.readAsDataURL(file); // 파일을 base64 text로 전환
-
-      reader.onload = () => {
-        baseURL = reader.result;
-        const sliceIndex = baseURL.indexOf(",");
-        setImagesrc((imagesrc = baseURL.substr(sliceIndex + 1)));
-        setImagePreview((prev) => [...prev, (imagePreview = baseURL)]);
-        resolve(baseURL);
-      };
-    });
-  };
-
-  // 이미지 확장자 index 반환 함수
-  const getExtension = (profilename) => {
-    if (profilename.indexOf("png") !== -1) {
-      return profilename.indexOf("png");
-    } else if (profilename.indexOf("jpg") !== -1) {
-      return profilename.indexOf("jpg");
-    }
-    if (profilename.indexOf("jpeg") !== -1) {
-      return profilename.indexOf("jpeg");
-    }
-  };
-
-  // 이미지 선택 함수
-  const onClickImage = (e) => {
-    const file = e.target.files[0];
-    if (file === undefined) {
-      e.preventDefault();
-      return;
-    }
-    const index = getExtension(file.name);
-    const imageFileRealName = file.name.substring(0, index - 1);
-    const imageFileExtension = file.type.split("/")[1];
-
-    if (roomImage.length < 5) {
-      setImage(file.name);
-      encodeFileToBase64(file).then(() => {
-        setRoomImage([
-          ...roomImage,
-          {
-            roomImageRealName: imageFileRealName,
-            roomImageExtension: imageFileExtension,
-            imageSource: imagesrc,
-          },
-        ]);
-      });
-    } else {
-      e.preventDefault();
-      alert("이미지는 최대 5개까지 설정할 수 있습니다!");
-    }
-  };
-
-  // 이미지 삭제 함수
-  const deleteImage = (e) => {
-    const selectedImageIndex = e.target.classList[1];
-    const imageIndex = selectedImageIndex.slice(4);
-    const selectedImg = document.querySelector(`.img-${imageIndex}`);
-
-    // input 비우기
-    setImage((image = ""));
-
-    // 프리뷰 삭제
-    setImagePreview((prev) =>
-      prev.filter((el) => {
-        return el !== selectedImg.src;
-      })
-    );
-
-    // 서버에게 전달할 객체배열 안에서 데이터 삭제
-    const sliceIndex = selectedImg.src.indexOf(",");
-    const src = selectedImg.src.substr(sliceIndex + 1);
-    setRoomImage((prev) =>
-      prev.filter((el) => {
-        return el.imageSource !== src;
-      })
-    );
-  };
-
-  // 태그 선택 함수
-  const onClickTag = (e) => {
-    if (e.target.classList[2] === "tag-clicked") {
-      e.target.classList.remove("tag-clicked");
-      setTag((prev) =>
-        prev.filter((el) => {
-          return el !== e.target.innerText;
-        })
-      );
-    } else {
-      if (tag.length > 9) {
-        e.preventDefault();
-        alert("태그는 최대 10개 선택 가능합니다!");
-      } else {
-        e.target.classList.add("tag-clicked");
-        setTag((prev) => [...prev, e.target.innerText]);
-      }
-    }
-  };
-
-  // test
-  useEffect(() => {
-    //console.log(roomContent);
-    console.log(roomImage);
-    console.log(imagePreview);
-    //console.log(tag);
-  });
-
   return (
     <>
       <div className="container">
@@ -186,92 +95,41 @@ const RoomTagInfo = () => {
             ></textarea>
           </div>
 
-          <div className="content-images">
-            <div className="images">
-              <p>사진추가</p>
-              <input readOnly className="image-name" value={image} />
-              <label htmlFor="filename">
-                <div>파일찾기</div>
-              </label>
-              <input
-                type="file"
-                id="filename"
-                accept=".jpg, .jpeg, .png"
-                onChange={onClickImage}
-              />
-            </div>
-
-            <div className="previews">
-              {imagePreview.map((preview, index) => {
-                return (
-                  <div className="preview" key={index}>
-                    <div>
-                      <img
-                        src={preview}
-                        alt="preview"
-                        className={`img-${index}`}
-                      />
-                    </div>
-                    <button className={`btn-${index}`} onClick={deleteImage}>
-                      X
-                    </button>
-                  </div>
-                );
-              })}
-            </div>
-          </div>
+          <SetRoomImages />
 
           <div className="content-tag">
             <p>빠른 태그 추가 (최대 10개)</p>
             <div className="tags">
               <div className="tags-age">
-                <div className="tag" onClick={onClickTag}>
-                  10대
-                </div>
-                <div className="tag" onClick={onClickTag}>
-                  20대
-                </div>
-                <div className="tag" onClick={onClickTag}>
-                  30대
-                </div>
-                <div className="tag" onClick={onClickTag}>
-                  40대
-                </div>
-                <div className="tag" onClick={onClickTag}>
-                  50대
-                </div>
-                <div className="tag" onClick={onClickTag}>
-                  60대
-                </div>
-                <div className="tag" onClick={onClickTag}>
-                  70대
-                </div>
+                {tagsAge.map((age, index) => {
+                  return (
+                    <div className="tag" onClick={onClickTag} key={index}>
+                      {age}
+                    </div>
+                  );
+                })}
                 <div className="tag-nomatter" onClick={onClickTag}>
                   연령 무관
                 </div>
               </div>
               <div className="tags-level-gender">
-                <div className="tag" onClick={onClickTag}>
-                  입문만
-                </div>
-                <div className="tag" onClick={onClickTag}>
-                  초보만
-                </div>
-                <div className="tag" onClick={onClickTag}>
-                  중수만
-                </div>
-                <div className="tag" onClick={onClickTag}>
-                  고수만
-                </div>
+                {tagsLevel.map((level, index) => {
+                  return (
+                    <div className="tag" onClick={onClickTag} key={index}>
+                      {level}
+                    </div>
+                  );
+                })}
                 <div className="tag-nomatter" onClick={onClickTag}>
                   실력 무관
                 </div>
-                <div className="tag" onClick={onClickTag}>
-                  남자만
-                </div>
-                <div className="tag" onClick={onClickTag}>
-                  여자만
-                </div>
+                {tagsGender.map((gender, index) => {
+                  return (
+                    <div className="tag" onClick={onClickTag} key={index}>
+                      {gender}
+                    </div>
+                  );
+                })}
                 <div className="tag-nomatter" onClick={onClickTag}>
                   성별 무관
                 </div>
@@ -331,107 +189,6 @@ const RoomTagInfo = () => {
           background-color: #f4f4f4;
           resize: none;
           font-size: 1.4em;
-        }
-
-        .content-images {
-          display: flex;
-          flex-direction: column;
-          align-items: center;
-          margin: 50px 0;
-        }
-
-        .images {
-          width: 550px;
-          height: 40px;
-          padding: 5px 10px 5px 14px;
-          margin-bottom: 20px;
-          border: solid 1px #e8e8e8;
-          border-radius: 10px;
-          display: flex;
-          align-items: center;
-          justify-content: space-between;
-        }
-
-        .images p {
-          font-size: 1.5em;
-          font-weight: bold;
-        }
-
-        .image-name {
-          width: 380px;
-          height: 30px;
-          border-style: none;
-          font-size: 1.4em;
-          padding: 5px;
-        }
-
-        .content-images label {
-          width: 70px;
-          height: 28px;
-          background-color: #08555f;
-          color: white;
-          font-size: 1.3em;
-          border: 0;
-          outline: 0;
-          cursor: pointer;
-          border-radius: 5px;
-          display: flex;
-          align-items: center;
-          justify-content: center;
-        }
-
-        .content-images input[type="file"] {
-          position: absolute;
-          width: 0;
-          height: 0;
-          padding: 0;
-          border: 0;
-          overflow: hidden;
-        }
-
-        .previews {
-          display: flex;
-          flex-direction: row;
-          justify-content: space-between;
-          align-items: center;
-        }
-
-        .preview {
-          display: table;
-          text-align: center;
-        }
-
-        .preview div {
-          width: 90px;
-          height: 90px;
-          margin: 0 10px;
-          border: none;
-          border-radius: 10px;
-          background-color: #d8d8d8;
-          overflow: hidden;
-          display: table-cell;
-          vertical-align: middle;
-        }
-
-        .preview img {
-          max-width: 90px;
-          max-height: 90px;
-        }
-
-        .preview button {
-          position: relative;
-          bottom: 10px;
-          right: 10px;
-          width: 20px;
-          height: 20px;
-          border: none;
-          border-radius: 50%;
-          opacity: 0.8;
-          background-color: #f2f2f2;
-          font-weight: bold;
-          font-size: 1.2em;
-          color: white;
-          cursor: pointer;
         }
 
         .content-tag {
